requirements.txt
Flask>=1.1.2
uwsgi>=2.0.19.1
kafka-python>=2.0.2
requests>=2.23.0
boto3>=1.0.0